Developer Index

This page lists the types and functions that are internal to the OAR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

OAR.DescretizedBNF

OAR.DescretizedBNF

OAR.inc_update_symbols!

OAR.iris_tt_real

OAR.is_nonterminal

OAR.is_terminal

OAR.join_gsymbol

OAR.new_weight

OAR.parse_grammar

OAR.process_statement!

OAR.random_produce

OAR.random_statement

OAR.real_to_symb

OAR.results_dir

OAR.symbolic_iris

OAR.trace!

OAR.update_dist!

OAR.update_node!

OAR.work_dir

Types

OAR.BNF

OAR.BNF

OAR.BNF

OAR.DataSplit

OAR.GramART

OAR.GramART

OAR.GramARTSymbol

OAR.Grammar

OAR.ProtoChildren

OAR.ProtoNodeStats

OAR.TerminalDist

OAR.VectoredDataSplit

OAR.VectoredDataSplit

Constants

Docs

Documentation for all internal names are listed below.

OAR.BNF — Type

struct BNF <: OAR.Grammar

Summary

Backus-Naur form of Grammar.

Consists of a set of terminal symbols, non-terminal symbols, and production rules.

Fields

N::Set{OAR.GSymbol}: Non-terminal symbols of the grammar.

T::Set{OAR.GSymbol}: Terminal symbols of the grammar.

S::Vector{OAR.GSymbol}: Definition of a statement in this grammar.

P::Dict{OAR.GSymbol, Set{OAR.GSymbol}}: The set of production rules of the grammar.

OAR.BNF — Method

BNF(S::Vector{OAR.GSymbol})

Summary

Constructor for a Backus-Naur Form grammer with an initial statement of non-terminal symbols.

Arguments

N::Statement: an initial set of non-terminal grammar symbols.

Method List / Definition Locations

BNF(S)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:121.

OAR.BNF — Method

BNF()

Summary

Default constructor for the Backus-Naur Form.

Method List / Definition Locations

BNF()

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:133.

OAR.DataSplit — Type

struct DataSplit

Summary

Train/test split dataset.

This struct contains a standardized train/test split of real-valued vectors of samples arranged in a matrix and mapping to integered labels.

Fields

train_x::Matrix{Float64}: The training data as a matrix of floating-point feature vectors: (n_features, n_samples).

test_x::Matrix{Float64}: The testing data as a matrix of floating-point feature vectors: (n_features, n_samples).

train_y::Vector{Int64}: The training labels as a vector of integer labels: (n_samples,).

test_y::Vector{Int64}: The testing labels as a vector of integer labels: (n_samples,)

OAR.GramART — Type

struct GramART

Summary

Fields

protonodes::ProtoNode: The OAR.ProtoNodes of the GramART module.

treenodes::Vector{TreeNode}: The OAR.TreeNodes of the GramART module.

grammar::OAR.BNF: The OAR.BNF-form grammar used for processing data (statements).

OAR.GramART — Method

GramART(grammar::OAR.BNF) -> OAR.GramART

Summary

Constructor for a OAR.GramART module that takes a BNF grammar and automatically sets up the ProtoNode tree.

Method List / Definition Locations

GramART(grammar)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:133.

OAR.GramARTSymbol — Type

GramARTSymbol

Description

Definition of Terminal symbols used throughtout GramART.

OAR.Grammar — Type

abstract type Grammar

Summary

Abstract type for formal grammars.

Fields

OAR.ProtoChildren — Type

ProtoChildren

Description

Alias for how ProtoNode children are indexed.

OAR.ProtoNodeStats — Method

ProtoNodeStats() -> OAR.ProtoNodeStats

Summary

Empty constructor for the mutable options and stats component of a ProtoNode.

Method List / Definition Locations

ProtoNodeStats()

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:164.

OAR.TerminalDist — Type

TerminalDist

Description

Terminal Distribution definition that is a dictionary mapping from Terminals to probabilities.

OAR.VectoredDataSplit — Type

struct VectoredDataSplit{T, M}

Summary

Vectored train/test split of arbitrary feature types.

This struct contains a standardized train/test split of vectors of vectored samples that map to labels.

Fields

train_x::Array{Vector{T}, 1} where T: Training data as a vector of feature vectors of type T.

test_x::Array{Vector{T}, 1} where T: Testing data as a vector of feature vectors of type T.

train_y::Vector: Training labels as a vector of type M.

test_y::Vector: Testing labels as a vector of type M.

OAR.VectoredDataSplit — Method

Convenience constructor, turning a OAR.DataSplit into its vectored variant.

Arguments

data::DataSplit: the original OAR.DataSplit to transform into a vectored data ssplit.

Base.show — Method

show(io::IO, bnf::OAR.BNF)

Summary

Overload of the show function for OAR.BNF.

Arguments

io::IO: the current IO stream.

bnf::BNF: the OAR.BNF grammar to print/display.

Method List / Definition Locations

show(io, bnf)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:148.

Base.show — Method

show(io::IO, data::OAR.DataSplit)

Summary

Overload of the show function for OAR.DataSplit.

Arguments

io::IO: the current IO stream.

data::DataSplit: the DataSplit to print/display.

Method List / Definition Locations

show(io, data)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:85.

Base.show — Method

show(io::IO, data::OAR.VectoredDataSplit)

Summary

Overload of the show function for OAR.VectoredDataSplit.

Arguments

io::IO: the current IO stream.

data::VectoredDataSplit: the OAR.VectoredDataSplit to print/display.

Method List / Definition Locations

show(io, data)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:103.

Base.show — Method

show(io::IO, node::ProtoNode)

Summary

Overload of the show function for OAR.ProtoNode.

Arguments

io::IO: the current IO stream.

node::ProtoNode: the OAR.ProtoNode to print/display.

Method List / Definition Locations

show(io, node)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:92.

OAR.DescretizedBNF — Method

DescretizedBNF(S::Vector{OAR.GSymbol}; bins) -> OAR.BNF

Summary

Creates a grammer for discretizing a set of symbols into a number of bins.

Arguments

N::Statement: the set of non-terminal grammar symbols to use for binning.

bins::Integer=10: optional, the granularity/number of bins.

Method List / Definition Locations

DescretizedBNF(S; bins)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:186.

OAR.DescretizedBNF — Method

DescretizedBNF(N::Vector{String}; bins) -> OAR.BNF

Summary

Wrapper for creating a DescretizedBNF from just a vector of nonterminal symbol names as strings.

This function turns the vector of strings in to a statement and passes it to the actual constructor.

Arguments

N::Vector{String}: the nonterminal symbol names as a vector of strings.

bins::Integer=10: optional, the granularity/number of bins.

Method List / Definition Locations

DescretizedBNF(N; bins)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:175.

OAR.inc_update_symbols! — Method

inc_update_symbols!(
    pn::ProtoNode,
    nonterminal::OAR.GSymbol{String},
    symb::OAR.GSymbol{String}
)

Summary

Updates the tree of protonodes from a single terminal.

Arguments

pn::ProtoNode: the top of the protonode tree to update.

nonterminal::GramARTSymbol: the nonterminal symbol of the statement to update at.

symb::GramARTSymbol: the terminal symbol to update everywhere.

Method List / Definition Locations

inc_update_symbols!(pn, nonterminal, symb)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:246.

OAR.iris_tt_real — Method

iris_tt_real() -> OAR.DataSplit

Summary

Loads the Iris dataset and returns a OAR.DataSplit.

Method List / Definition Locations

iris_tt_real()

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:138.

OAR.is_nonterminal — Method

is_nonterminal(
    grammar::OAR.Grammar,
    symb::OAR.GSymbol
) -> Any

Summary

Checks if a symbol is non-terminal in the grammar.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

symb::GSymbol: the grammar symbol to use.

Method List / Definition Locations

is_nonterminal(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:263.

OAR.is_terminal — Method

is_terminal(grammar::OAR.Grammar, symb::OAR.GSymbol) -> Any

Summary

Checks if a symbol is terminal in the grammar.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

symb::GSymbol: the grammar symbol to use.

Method List / Definition Locations

is_terminal(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:253.

OAR.join_gsymbol — Method

join_gsymbol(
    symb::OAR.GSymbol,
    num::Integer;
    terminal
) -> OAR.GSymbol{String}

Summary

Returns a new GSymbol by adding a suffix.

Method List / Definition Locations

join_gsymbol(symb, num; terminal)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:159.

OAR.new_weight — Method

new_weight(
    dist::Dict{OAR.GSymbol{String}, Float64},
    N::Dict{OAR.GSymbol{String}, Int64}
)

Summary

Method List / Definition Locations

new_weight(dist, N)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:284.

OAR.parse_grammar — Method

parse_grammar(
    grammar::OAR.Grammar,
    statement::Vector{OAR.GSymbol}
)

Summary

Parses and checks that a statement is permissible under a grammer.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

statement::Statement: a grammar [OAR.Statement] to check the validity of.s

Method List / Definition Locations

parse_grammar(grammar, statement)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:233.

OAR.process_statement! — Method

process_statement!(
    gramart::OAR.GramART,
    statement::Vector{OAR.GSymbol}
)

Summary

Processes a statement for a OAR.GramART module.

Arguments

gramart::GramART: the OAR.GramART to update with the statement.

statement::Statement: the grammar statement to process.

Method List / Definition Locations

process_statement!(gramart, statement)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:267.

OAR.random_produce — Method

random_produce(
    grammar::OAR.Grammar,
    symb::OAR.GSymbol
) -> Any

Summary

Produces a random terminal from the non-terminal using the corresponding production rule.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

symb::GSymbol: the grammar symbol to use.

Method List / Definition Locations

random_produce(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:243.

OAR.random_statement — Method

random_statement(
    grammar::OAR.Grammar
) -> Vector{OAR.GSymbol}

Summary

Generates a random statement from a grammar.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

Method List / Definition Locations

random_statement(grammar)

defined at /home/runner/work/OAR/OAR/src/lib/BNF.jl:274.

OAR.real_to_symb — Method

real_to_symb(
    data::OAR.DataSplit,
    labels::Vector{String};
    bins
) -> Tuple{OAR.VectoredDataSplit{OAR.GSymbol, Int64}, OAR.BNF}

Summary

Turns a OAR.DataSplit into a binned symbolic variant for use with GramART.

Arguments

data::DataSplit: the OAR.DataSplit to convert to symbols.

labels::Vector{String}: the labels corresponding to the non-terminal symbol names for the feature categories and their subsequent terminal variants.

bins::Int=10: optional, the number of symbols to descretize the real-valued data to.

Method List / Definition Locations

real_to_symb(data, labels; bins)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:166.

OAR.results_dir — Method

results_dir(args...) -> String

Summary

Points to the results directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

results_dir(args)

defined at /home/runner/work/OAR/OAR/src/lib/drwatson.jl:40.

OAR.symbolic_iris — Method

symbolic_iris(
;
    bins
) -> Tuple{OAR.VectoredDataSplit{OAR.GSymbol, Int64}, OAR.BNF}

Summary

Quickly generates a [OAR.VectoredDataSplit] of the symbolic Iris dataset.

Arguments

bins::Int=10: optional, the number of symbols to descretize the real-valued data to.

Method List / Definition Locations

symbolic_iris(; bins)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:258.

OAR.trace! — Method

trace!(
    A::TreeNode,
    B::ProtoNode,
    sum::AbstractFloat,
    size::Integer
)

Summary

Method List / Definition Locations

trace!(A, B, sum, size)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:275.

OAR.update_dist! — Method

update_dist!(pn::ProtoNode, symb::OAR.GSymbol{String})

Summary

Updates the distribution of a single OAR.ProtoNode from one new symbol instance.

Arguments

pn::ProtoNode: the OAR.ProtoNode to update the distribution with.

symb::GramARTSymbol: the symbol instance to update the OAR.ProtoNode with.

Method List / Definition Locations

update_dist!(pn, symb)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:226.

OAR.update_node! — Method

update_node!(A::TreeNode, B::ProtoNode)

Summary

Method List / Definition Locations

update_node!(A, B)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:291.

OAR.work_dir — Method

work_dir(args...) -> String

Summary

Points to the work directory containing raw datasets, processed datasets, and results.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

work_dir(args)

defined at /home/runner/work/OAR/OAR/src/lib/drwatson.jl:29.