Index

This page lists the core functions and types of the Julia component of the OAR project.

"Table" of Contents

Modules	Functions	Types	Constants

Index	Index	Index	Index

Docs	Docs	Docs	Docs

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

OAR.OAR

Functions

OAR.classify

OAR.classify

OAR.classify

OAR.train!

OAR.train!

OAR.train!

Types

OAR.DDVSTART

OAR.DDVSTART

OAR.DDVSTART

OAR.DVSTART

OAR.DVSTART

OAR.DVSTART

OAR.GSymbol

OAR.GSymbol

OAR.ProtoNode

OAR.ProtoNode

OAR.ProtoNode

OAR.START

OAR.START

OAR.START

OAR.TreeNode

OAR.TreeNode

OAR.TreeNode

Constants

OAR.OAR_VERSION

Docs

This section lists the documentation for every exported name of the OAR package.

Modules

OAR.OAR — Module

A module encapsulating the experimental driver code for the OAR project.

Imports

The following names are imported by the package as dependencies:

AdaptiveResonance

ArgParse

Base

ClusterValidityIndices

Clustering

Core

DataFrames

DelimitedFiles

Distributed

DocStringExtensions

DrWatson

InvertedIndices

Latexify

Lerche

MLDatasets

MLUtils

NumericalTypeAliases

Pkg

Plots

ProgressMeter

Random

StatsPlots

Exports

The following names are exported and available when using the package:

DDVSTART

DVSTART

GSymbol

OAR_VERSION

ProtoNode

START

TreeNode

classify

train!

Functions

OAR.classify — Method

classify(
    art::DDVSTART,
    statement::Union{TreeNode, Array{GSymbol{T}, 1} where T};
    get_bmu
) -> Int64

Summary

Classifies SomeStatement with the provided DDVSTART module with an optional flag to get the best-matching unit in the case of a complete mismatch.

Arguments

art::DDVSTART: the DDVSTART module to classify with.

statement::SomeStatment: the sample as SomeStatment to classify.

get_bmu: optional flag to get the best-matching unit in the case of complete mismatch.

Method List / Definition Locations

classify(art, statement; get_bmu)

defined at /home/runner/work/OAR/OAR/src/lib/start/ddv.jl:440.

OAR.classify — Method

classify(
    art::DVSTART,
    statement::Array{GSymbol{T}, 1} where T;
    get_bmu
) -> Int64

Summary

Classifies the OAR.Statement into one of OAR.DVSTART's internal categories.

Arguments

art::DVSTART: the OAR.DVSTART to use in classification/inference.

statement::Statement: the OAR.Statement to classify.

get_bmu::Bool=false: optional, whether to get the best matching unit in the case of complete mismatch.

Method List / Definition Locations

classify(art, statement; get_bmu)

defined at /home/runner/work/OAR/OAR/src/lib/start/dv.jl:219.

OAR.classify — Method

classify(
    art::START,
    statement::Union{TreeNode, Array{GSymbol{T}, 1} where T};
    get_bmu
) -> Int64

Summary

Classifies the OAR.Statement into one of OAR.START's internal categories.

Arguments

art::START: the OAR.START to use in classification/inference.

statement::SomeStatement: the OAR.SomeStatement to classify.

get_bmu::Bool=false: optional, whether to get the best matching unit in the case of complete mismatch.

Method List / Definition Locations

classify(art, statement; get_bmu)

defined at /home/runner/work/OAR/OAR/src/lib/start/start.jl:224.

OAR.train! — Method

train!(
    art::DDVSTART,
    statement::Union{TreeNode, Array{GSymbol{T}, 1} where T};
    y
) -> Int64

Summary

Trains a DDVSTART module on SomeStatement and an optional supervisory label.

Arguments

art::DDVSTART: the DDVSTART module to train.

statement::SomeStatement: the statement sample to train upon.

y::Integer=0: optional supervisory label for the sample.

Method List / Definition Locations

train!(art, statement; y)

defined at /home/runner/work/OAR/OAR/src/lib/start/ddv.jl:365.

OAR.train! — Method

train!(
    art::DVSTART,
    statement::Union{TreeNode, Array{GSymbol{T}, 1} where T};
    y
) -> Int64

Summary

Trains OAR.DVSTART module on a OAR.SomeStatement from the OAR.DVSTART's grammar.

Arguments

art::DVSTART: the OAR.DVSTART to update with the OAR.SomeStatement.

statement::SomeStatement: the grammar OAR.SomeStatement to process.

y::Integer=0: optional supervised label as an integer.

Method List / Definition Locations

train!(art, statement; y)

defined at /home/runner/work/OAR/OAR/src/lib/start/dv.jl:147.

OAR.train! — Method

train!(
    art::START,
    statement::Union{TreeNode, Array{GSymbol{T}, 1} where T};
    y
) -> Int64

Summary

Trains OAR.START module on a OAR.SomeStatement from the OAR.START's grammar.

Arguments

art::START: the OAR.START to update with the OAR.SomeStatement.

statement::SomeStatement: the grammar OAR.SomeStatement to train upon.

y::Integer=0: optional supervised label as an integer.

Method List / Definition Locations

train!(art, statement; y)

defined at /home/runner/work/OAR/OAR/src/lib/start/start.jl:160.

Types

OAR.DDVSTART — Type

mutable struct DDVSTART <: OAR.AbstractSTART

Summary

A DDVSTART module.

Fields

grammar::OAR.CFG: The OAR.CFG (Context-Free Grammar) used for processing data (statements).

opts::OAR.opts_DDVSTART: DDVSTART options struct.

subopts::OAR.opts_START: START options struct used for all F2 nodes.

F2::Vector{START}: List of F2 nodes (themselves START modules).

labels::Vector{Int64}: Incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

n_categories::Int64: Number of total categories.

epoch::Int64: Current training epoch.

T::Vector{Float64}: DDVFA activation values.

M::Vector{Float64}: DDVFA match values.

stats::Dict{String, Any}: Dictionary of mutable statistics for the module.

OAR.DDVSTART — Method

DDVSTART(
    grammar::OAR.CFG,
    opts::OAR.opts_DDVSTART
) -> DDVSTART

Summary

DDVSTART constructor taking a CFG grammart and opts_DDVSTART options struct.

Arguments

grammar:CFG: the grammar that the module works upon.

opts::opts_DDVSTART: the options of

Method List / Definition Locations

DDVSTART(grammar, opts)

defined at /home/runner/work/OAR/OAR/src/lib/start/ddv.jl:222.

OAR.DDVSTART — Method

DDVSTART(grammar::OAR.CFG; kwargs...) -> DDVSTART

Summary

DDVSTART constructor taking a CFG grammart and keyword arguments for its options.

Arguments

grammar:CFG: the grammar that the module works upon.

Method List / Definition Locations

DDVSTART(grammar; kwargs...)

defined at /home/runner/work/OAR/OAR/src/lib/start/ddv.jl:201.

OAR.DVSTART — Type

struct DVSTART <: OAR.SingleSTART

Summary

Definition of a DVSTART module.

Contains the ProtoNodes and CFG grammar that is used for processing statements and generating nodes.

Fields

protonodes::Vector{ProtoNode}: The OAR.ProtoNodes of the DVSTART module.

grammar::OAR.CFG: The OAR.CFG (Context-Free Grammar) used for processing data (statements).

opts::OAR.opts_DVSTART: The OAR.opts_DVSTART hyperparameters of the DVSTART module.

labels::Vector{Int64}: Incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

T::Vector{Float64}: Activation values for every weight for a given sample.

M::Vector{Float64}: Match values for every weight for a given sample.

stats::Dict{String, Any}: Dictionary of mutable statistics for the module.

OAR.DVSTART — Method

DVSTART(grammar::OAR.CFG, opts::OAR.opts_DVSTART) -> DVSTART

Summary

Constructor for an OAR.DVSTART module that takes a CFG grammar and automatically sets up the ProtoNode tree.

Arguments

grammar::CFG: the [OAR.CFG] context-free grammar to use.

opts::opts_DVSTART: a custom set of OAR.DVSTART options to use.

Method List / Definition Locations

DVSTART(grammar, opts)

defined at /home/runner/work/OAR/OAR/src/lib/start/dv.jl:98.

OAR.DVSTART — Method

DVSTART(grammar::OAR.CFG; kwargs...) -> DVSTART

Summary

Constructor for an OAR.DVSTART module that takes a OAR.CFG grammar and an optional list of keyword arguments for the options.

Arguments

grammar::CFG: the [OAR.CFG] context-free grammar to use.

kwargs...: a list of keyword arguments for the OAR.opts_DVSTART options struct.

Method List / Definition Locations

DVSTART(grammar; kwargs...)

defined at /home/runner/work/OAR/OAR/src/lib/start/dv.jl:124.

OAR.GSymbol — Type

struct GSymbol{T}

Summary

Definition of a grammar symbol with arbitrary datatype and a boolean flag for if the symbol is terminal or not.

Fields

data::Any: The grammar symbol of type T.

terminal::Bool: Boolean flag if the symbol is terminal (true) or nonterminal (false).

OAR.GSymbol — Method

Constructor for a grammar symbol from just the provided data (defaults to being terminal).

Arguments

data::T where T <: Any: the piece of data comprising the grammar symbol of any type.

OAR.ProtoNode — Type

struct ProtoNode <: OAR.ARTNode

Summary

ProtoNode struct, used to generate tree prototypes, which are the templates of OAR.START.

Fields

dist::Dict{GSymbol{String}, Float64}: The TerminalDist distribution over all symbols at this node.

N::Dict{GSymbol{String}, Int64}: The SymbolCount update counters for each symbol.

children::Dict{GSymbol{String}, ProtoNode}: The children of this node (Dict{STARTSymbol, ProtoNode}).

stats::OAR.ProtoNodeStats: The mutable ProtoNodeStats options and stats of the ProtoNode.

OAR.ProtoNode — Method

ProtoNode(symbols::Set{GSymbol{T}} where T) -> ProtoNode

Summary

Constructor for a zero-initialized OAR.START OAR.ProtoNode.

Arguments

symbols::SymbolSet: the terminal symbols to initialize the node with.

Method List / Definition Locations

ProtoNode(symbols)

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:179.

OAR.ProtoNode — Method

ProtoNode() -> ProtoNode

Summary

Empty constructor for a OAR.START OAR.ProtoNode.

Method List / Definition Locations

ProtoNode()

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:163.

OAR.START — Type

struct START <: OAR.SingleSTART

Summary

Definition of a START module.

Contains the ProtoNodes and CFG grammar that is used for processing statements and generating nodes.

Fields

protonodes::Vector{ProtoNode}: The OAR.ProtoNodes of the START module.

grammar::OAR.CFG: The OAR.CFG (Context-Free Grammar) used for processing data (statements).

opts::OAR.opts_START: The OAR.opts_START hyperparameters of the START module.

labels::Vector{Int64}: Incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

T::Vector{Float64}: Activation values for every weight for a given sample.

M::Vector{Float64}: Match values for every weight for a given sample.

stats::Dict{String, Any}: Dictionary of mutable statistics for the module.

OAR.START — Method

START(grammar::OAR.CFG, opts::OAR.opts_START) -> START

Summary

Constructor for an OAR.START module that takes a CFG grammar and automatically sets up the ProtoNode tree.

Arguments

grammar::CFG: the [OAR.CFG] context-free grammar to use.

opts::opts_START: a custom set of OAR.START options to use.

Method List / Definition Locations

START(grammar, opts)

defined at /home/runner/work/OAR/OAR/src/lib/start/start.jl:108.

OAR.START — Method

START(grammar::OAR.CFG; kwargs...) -> START

Summary

Constructor for an OAR.START module that takes a OAR.CFG grammar and an optional list of keyword arguments for the options.

Arguments

grammar::CFG: the [OAR.CFG] context-free grammar to use.

kwargs...: a list of keyword arguments for the OAR.opts_START options struct.

Method List / Definition Locations

START(grammar; kwargs...)

defined at /home/runner/work/OAR/OAR/src/lib/start/start.jl:134.

OAR.TreeNode — Type

struct TreeNode <: OAR.ARTNode

Summary

Tree node for a START module.

Fields

t::GSymbol{String}: The STARTSymbol symbol for the node.

children::Vector{TreeNode}: Children nodes of this node.

OAR.TreeNode — Type

TreeNode(name::AbstractString) -> TreeNode
TreeNode(
    name::AbstractString,
    is_terminal::Bool
) -> TreeNode

Summary

Constructor for a OAR.START OAR.TreeNode, taking a string name of the symbol and if it is terminal or not.

Arguments

name::AbstractString: the string name of the symbol to instantiate the OAR.TreeNode with.

is_terminal::Bool: flag for if the symbol in the node is terminal or not.

Method List / Definition Locations

TreeNode(name)
TreeNode(name, is_terminal)

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:213.

OAR.TreeNode — Method

TreeNode(symb::GSymbol{String}) -> TreeNode

Summary

Constructor for a OAR.START OAR.TreeNode taking an existing OAR.STARTSymbol.

Arguments

symb::STARTSymbol: the preconstructed STARTSymbol used for constructing the OAR.TreeNode.

Method List / Definition Locations

TreeNode(symb)

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:199.

Constants

OAR.OAR_VERSION — Constant

A constant that contains the version of the installed OAR package.

This value is computed at compile time, so it may be used to programmatically verify the version of OAR that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.