Developer Index

This page lists the types and functions that are internal to the OAR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

OAR.DescretizedCFG

OAR.DescretizedCFG

OAR.NonTerminal

OAR.Terminal

OAR.add_node!

OAR.data_dir

OAR.inc_update_symbols!

OAR.iris_tt_real

OAR.is_nonterminal

OAR.is_terminal

OAR.join_gsymbol

OAR.new_weight

OAR.parse_grammar

OAR.process_statement!

OAR.quick_statement

OAR.random_produce

OAR.random_statement

OAR.real_to_symb

OAR.results_dir

OAR.symbolic_iris

OAR.trace!

OAR.train!

OAR.update_dist!

OAR.update_node!

OAR.work_dir

Types

OAR.ARTNode

OAR.CFG

OAR.CFG

OAR.CFG

OAR.DataSplit

OAR.GSymbol

OAR.GSymbol

OAR.GramART

OAR.GramART

OAR.GramARTSymbol

OAR.Grammar

OAR.ProductionRule

OAR.ProductionRuleSet

OAR.ProtoChildren

OAR.ProtoNodeStats

OAR.ProtoNodeStats

OAR.Statement

OAR.SymbolSet

OAR.TerminalDist

OAR.VectoredDataSplit

OAR.VectoredDataSplit

Constants

OAR.GRAMMAR_ARG

OAR.GRAMMAR_SYMB_ARG

OAR.GSYMBOL_DATA_ARG

Docs

Documentation for all internal names are listed below.

OAR.GRAMMAR_ARG — Constant

GRAMMAR_ARG

Description

Common docstring argument for grammars.

OAR.GRAMMAR_SYMB_ARG — Constant

GRAMMARSYMBARG

Description

Common docstring for functions using a grammar and a grammar symbol.

OAR.GSYMBOL_DATA_ARG — Constant

GSYMBOLDATAARG

Description

Common argument docstring for GSymbol consruction.

OAR.ARTNode — Type

abstract type ARTNode

Summary

Definition of the ARTNode supertype.

Fields

OAR.CFG — Type

struct CFG <: OAR.Grammar

Summary

Context-Free Grammar.

Consists of a set of terminal symbols, non-terminal symbols, and production rules of Backus-Naur Form.

Fields

N::Set{OAR.GSymbol}: Non-terminal symbols of the grammar.

T::Set{OAR.GSymbol}: Terminal symbols of the grammar.

S::Vector{OAR.GSymbol}: Definition of a statement in this grammar.

P::Dict{OAR.GSymbol, Set{OAR.GSymbol}}: The set of production rules of the grammar of the Backus-Naur Form (CFG).

OAR.CFG — Method

CFG(S::Vector{OAR.GSymbol})

Summary

Constructor for a Context-Free Grammer with an initial statement of non-terminal symbols.

Arguments

N::Statement: an initial set of non-terminal grammar symbols.

Method List / Definition Locations

CFG(S)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:87.

OAR.CFG — Method

CFG()

Summary

Default constructor for a Context-Free Grammar.

Method List / Definition Locations

CFG()

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:99.

OAR.DataSplit — Type

struct DataSplit

Summary

Train/test split dataset.

This struct contains a standardized train/test split of real-valued vectors of samples arranged in a matrix and mapping to integered labels.

Fields

train_x::Matrix{Float64}: The training data as a matrix of floating-point feature vectors: (n_features, n_samples).

test_x::Matrix{Float64}: The testing data as a matrix of floating-point feature vectors: (n_features, n_samples).

train_y::Vector{Int64}: The training labels as a vector of integer labels: (n_samples,).

test_y::Vector{Int64}: The testing labels as a vector of integer labels: (n_samples,)

OAR.GSymbol — Type

struct GSymbol{T}

Summary

Definition of a grammar symbol with arbitrary datatype and a boolean flag for if the symbol is terminal or not.

Fields

data::Any: The grammar symbol of type T.

terminal::Bool: Boolean flag if the symbol is terminal (true) or nonterminal (false).

OAR.GSymbol — Method

Constructor for a grammar symbol from just the provided data (defaults to being terminal).

Arguments

data::T where T <: Any: the piece of data comprising the grammar symbol of any type.

OAR.GramART — Type

struct GramART

Summary

Definition of a GramART module.

Contains the proto nodes, tree nodes, and grammar that is used.

Fields

protonodes::Vector{ProtoNode}: The OAR.ProtoNodes of the GramART module.

treenodes::Vector{TreeNode}: The OAR.TreeNodes of the GramART module.

grammar::OAR.CFG: The OAR.CFG (Context-Free Grammar) used for processing data (statements).

OAR.GramART — Method

GramART(grammar::OAR.CFG) -> OAR.GramART

Summary

Constructor for a OAR.GramART module that takes a CFG grammar and automatically sets up the ProtoNode tree.

Method List / Definition Locations

GramART(grammar)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:144.

OAR.GramARTSymbol — Type

GramARTSymbol

Description

Definition of Terminal symbols used throughtout GramART.

OAR.Grammar — Type

abstract type Grammar

Summary

Abstract type for formal grammars.

Fields

OAR.ProductionRule — Type

ProductionRule

Description

Type alias (ProductionRule = [OAR.SymbolSet](@ref)), a grammar production rule is a set of symbols.

OAR.ProductionRuleSet — Type

ProductionRuleSet

Description

Type alias (ProductionRuleSet = Dict{[OAR.GSymbol](@ref), [OAR.ProductionRule](@ref)}), a production rule set is a dictionary mapping grammar symbols to production rules.

OAR.ProtoChildren — Type

ProtoChildren

Description

Alias for how ProtoNode children are indexed.

OAR.ProtoNodeStats — Type

mutable struct ProtoNodeStats

Summary

The mutable components of a ProtoNode, containing options and statistics of the node.

Fields

m::Int64: Convenience counter for the total number of symbols encountered.

terminal::Bool: If the ProtoNode is terminal on the graph.

OAR.ProtoNodeStats — Method

ProtoNodeStats() -> OAR.ProtoNodeStats

Summary

Empty constructor for the mutable options and stats component of a ProtoNode.

Method List / Definition Locations

ProtoNodeStats()

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:172.

OAR.Statement — Type

Statement

Description

Type alias (Statement = Vector{[OAR.GSymbol](@ref)}), a statement is a vector of grammar symbols.

OAR.SymbolSet — Type

SymbolSet

Description

Type alias (SymbolSet = Set{[OAR.GSymbol](@ref)}), a set of grammar symbols is implemented as a Julia set.

OAR.TerminalDist — Type

TerminalDist

Description

Terminal Distribution definition that is a dictionary mapping from Terminals to probabilities.

OAR.VectoredDataSplit — Type

struct VectoredDataSplit{T, M}

Summary

Vectored train/test split of arbitrary feature types.

This struct contains a standardized train/test split of vectors of vectored samples that map to labels.

Fields

train_x::Array{Vector{T}, 1} where T: Training data as a vector of feature vectors of type T.

test_x::Array{Vector{T}, 1} where T: Testing data as a vector of feature vectors of type T.

train_y::Vector: Training labels as a vector of type M.

test_y::Vector: Testing labels as a vector of type M.

OAR.VectoredDataSplit — Method

Convenience constructor, turning a OAR.DataSplit into its vectored variant.

Arguments

data::DataSplit: the original OAR.DataSplit to transform into a vectored data ssplit.

Base.show — Method

show(io::IO, cfg::OAR.CFG)

Summary

Overload of the show function for OAR.CFG.

Arguments

io::IO: the current IO stream.

cfg::CFG: the OAR.CFG OAR.Grammar to print/display.

Method List / Definition Locations

show(io, cfg)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:127.

Base.show — Method

show(io::IO, data::OAR.DataSplit)

Summary

Overload of the show function for OAR.DataSplit.

Arguments

io::IO: the current IO stream.

data::DataSplit: the OAR.DataSplit to print/display.

Method List / Definition Locations

show(io, data)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:85.

Base.show — Method

show(io::IO, data::OAR.VectoredDataSplit)

Summary

Overload of the show function for OAR.VectoredDataSplit.

Arguments

io::IO: the current IO stream.

data::VectoredDataSplit: the OAR.VectoredDataSplit to print/display.

Method List / Definition Locations

show(io, data)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:103.

Base.show — Method

show(io::IO, node::ProtoNode)

Summary

Overload of the show function for OAR.ProtoNode.

Arguments

io::IO: the current IO stream.

node::ProtoNode: the OAR.ProtoNode to print/display.

Method List / Definition Locations

show(io, node)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:265.

OAR.DescretizedCFG — Method

DescretizedCFG(S::Vector{OAR.GSymbol}; bins) -> OAR.CFG

Summary

Creates a grammer for discretizing a set of symbols into a number of bins.

Arguments

N::Statement: the set of non-terminal grammar symbols to use for binning.

bins::Integer=10: optional, the granularity/number of bins.

Method List / Definition Locations

DescretizedCFG(S; bins)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:155.

OAR.DescretizedCFG — Method

DescretizedCFG(N::Vector{String}; bins) -> OAR.CFG

Summary

Wrapper for creating a DescretizedCFG from just a vector of nonterminal symbol names as strings.

This function turns the vector of strings in to a statement and passes it to the actual constructor.

Arguments

N::Vector{String}: the nonterminal symbol names as a vector of strings.

bins::Integer=10: optional, the granularity/number of bins.

Method List / Definition Locations

DescretizedCFG(N; bins)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:144.

OAR.NonTerminal — Method

NonTerminal(data) -> OAR.GSymbol

Summary

Convenience consructor for a nonterminal grammar symbol.

Arguments

data::T where T <: Any: the piece of data comprising the grammar symbol of any type.

Method List / Definition Locations

NonTerminal(data)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/symbols.jl:68.

OAR.Terminal — Method

Terminal(data) -> OAR.GSymbol

Summary

Convenience constructor for a terminal grammar symbol.

Arguments

data::T where T <: Any: the piece of data comprising the grammar symbol of any type.

Method List / Definition Locations

Terminal(data)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/symbols.jl:56.

OAR.add_node! — Method

add_node!(gramart::OAR.GramART) -> Vector{ProtoNode}

Summary

Adds a recursively-generated OAR.ProtoNode to the OAR.GramART module.

Arguments

gramart::GramART: the OAR.GramART to append a new node to.

Method List / Definition Locations

add_node!(gramart)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:233.

OAR.data_dir — Method

data_dir(args...) -> String

Summary

Points to the data directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

data_dir(args)

defined at /home/runner/work/OAR/OAR/src/lib/drwatson.jl:51.

OAR.inc_update_symbols! — Method

inc_update_symbols!(
    pn::ProtoNode,
    nonterminal::OAR.GSymbol{String},
    symb::OAR.GSymbol{String}
)

Summary

Updates the tree of protonodes from a single terminal.

Arguments

pn::ProtoNode: the top of the protonode tree to update.

nonterminal::GramARTSymbol: the nonterminal symbol of the statement to update at.

symb::GramARTSymbol: the terminal symbol to update everywhere.

Method List / Definition Locations

inc_update_symbols!(pn, nonterminal, symb)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:296.

OAR.iris_tt_real — Method

iris_tt_real(; download_local) -> OAR.DataSplit

Summary

Loads the Iris dataset and returns a OAR.DataSplit.

Arguments

download_local::Bool=false: optional (default false), to download the Iris dataset to the local datadir.

Method List / Definition Locations

iris_tt_real(; download_local)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:141.

OAR.is_nonterminal — Method

is_nonterminal(
    grammar::OAR.Grammar,
    symb::OAR.GSymbol
) -> Any

Summary

Checks if a symbol is non-terminal in the grammar.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

symb::GSymbol: the grammar symbol to use.

Method List / Definition Locations

is_nonterminal(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:238.

OAR.is_terminal — Method

is_terminal(grammar::OAR.Grammar, symb::OAR.GSymbol) -> Any

Summary

Checks if a symbol is terminal in the grammar.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

symb::GSymbol: the grammar symbol to use.

Method List / Definition Locations

is_terminal(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:228.

OAR.join_gsymbol — Method

join_gsymbol(
    symb::OAR.GSymbol,
    num::Integer;
    terminal
) -> OAR.GSymbol{String}

Summary

Returns a new OAR.GSymbol by adding a suffix.

Arguments

symb::GSymbol: the original symbol to append a suffix to.

num::Integer: the integer to add as a suffix to the symbol.

terminal::Bool=true: optional (default true), to set the new symbol as terminal.

Method List / Definition Locations

join_gsymbol(symb, num; terminal)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/symbols.jl:87.

OAR.new_weight — Method

new_weight(
    dist::Dict{OAR.GSymbol{String}, Float64},
    N::Dict{OAR.GSymbol{String}, Int64}
)

Summary

Method List / Definition Locations

new_weight(dist, N)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:353.

OAR.parse_grammar — Method

parse_grammar(
    grammar::OAR.Grammar,
    statement::Vector{OAR.GSymbol}
)

Summary

Parses and checks that a statement is permissible under a grammer.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

statement::Statement: a grammar [OAR.Statement] to check the validity of.s

Method List / Definition Locations

parse_grammar(grammar, statement)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:208.

OAR.process_statement! — Method

process_statement!(
    gramart::OAR.GramART,
    statement::Vector{OAR.GSymbol},
    index::Int64
)

Summary

Processes a statement for a OAR.GramART module.

Arguments

gramart::GramART: the OAR.GramART to update with the statement.

statement::Statement: the grammar statement to process.

Method List / Definition Locations

process_statement!(gramart, statement, index)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:317.

OAR.quick_statement — Method

quick_statement(data::Array{T, 1}; terminal) -> Any

Summary

Creates a OAR.Statement from a vector of elements of arbitrary type.

Arguments

data::Vector{T} where T<:Any: a vector of any type for creating a OAR.Statement of symbols of that type.

terminal::Bool=false: optional, if the symbols of the statement are terminal.

Method List / Definition Locations

quick_statement(data; terminal)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:114.

OAR.random_produce — Method

random_produce(
    grammar::OAR.Grammar,
    symb::OAR.GSymbol
) -> Any

Summary

Produces a random terminal from the non-terminal using the corresponding production rule.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

symb::GSymbol: the grammar symbol to use.

Method List / Definition Locations

random_produce(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:218.

OAR.random_statement — Method

random_statement(
    grammar::OAR.Grammar
) -> Vector{OAR.GSymbol}

Summary

Generates a random statement from a grammar.

Arguments

grammar::Grammar: a subtype of the abstract OAR.Grammar type.

Method List / Definition Locations

random_statement(grammar)

defined at /home/runner/work/OAR/OAR/src/lib/grammar/grammar.jl:249.

OAR.real_to_symb — Method

real_to_symb(
    data::OAR.DataSplit,
    labels::Vector{String};
    bins
) -> Tuple{OAR.VectoredDataSplit{OAR.GSymbol, Int64}, OAR.CFG}

Summary

Turns a OAR.DataSplit into a binned symbolic variant for use with GramART.

Arguments

data::DataSplit: the OAR.DataSplit to convert to symbols.

labels::Vector{String}: the labels corresponding to the non-terminal symbol names for the feature categories and their subsequent terminal variants.

bins::Int=10: optional, the number of symbols to descretize the real-valued data to.

Method List / Definition Locations

real_to_symb(data, labels; bins)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:174.

OAR.results_dir — Method

results_dir(args...) -> String

Summary

Points to the results directory.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

results_dir(args)

defined at /home/runner/work/OAR/OAR/src/lib/drwatson.jl:40.

OAR.symbolic_iris — Method

symbolic_iris(
;
    bins,
    download_local
) -> Tuple{OAR.VectoredDataSplit{OAR.GSymbol, Int64}, OAR.CFG}

Summary

Quickly generates a [OAR.VectoredDataSplit] of the symbolic Iris dataset.

Arguments

bins::Int=10: optional, the number of symbols to descretize the real-valued data to.

download_local::Bool=false: optional (default false), to download the Iris dataset to the local datadir.

Method List / Definition Locations

symbolic_iris(; bins, download_local)

defined at /home/runner/work/OAR/OAR/src/lib/data_utils.jl:267.

OAR.trace! — Method

trace!(
    A::TreeNode,
    B::ProtoNode,
    sum::AbstractFloat,
    size::Integer
)

Summary

Method List / Definition Locations

trace!(A, B, sum, size)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:344.

OAR.train! — Method

train!(gramart::OAR.GramART, statement::Vector{OAR.GSymbol})

Summary

Processes a statement for a OAR.GramART module.

Arguments

gramart::GramART: the OAR.GramART to update with the statement.

statement::Statement: the grammar statement to process.

Method List / Definition Locations

train!(gramart, statement)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:330.

OAR.update_dist! — Method

update_dist!(pn::ProtoNode, symb::OAR.GSymbol{String})

Summary

Updates the distribution of a single OAR.ProtoNode from one new symbol instance.

Arguments

pn::ProtoNode: the OAR.ProtoNode to update the distribution with.

symb::GramARTSymbol: the symbol instance to update the OAR.ProtoNode with.

Method List / Definition Locations

update_dist!(pn, symb)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:276.

OAR.update_node! — Method

update_node!(A::TreeNode, B::ProtoNode)

Summary

Method List / Definition Locations

update_node!(A, B)

defined at /home/runner/work/OAR/OAR/src/lib/GramART.jl:360.

OAR.work_dir — Method

work_dir(args...) -> String

Summary

Points to the work directory containing raw datasets, processed datasets, and results.

Arguments

args...: the string directories to append to the directory.

Method List / Definition Locations

work_dir(args)

defined at /home/runner/work/OAR/OAR/src/lib/drwatson.jl:29.