Developer Index

This page lists the types and functions that are internal to the OAR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

OAR.DescretizedBNF

OAR.is_nonterminal

OAR.is_terminal

OAR.join_gsymbol

OAR.parse_grammar

OAR.random_produce

OAR.random_statement

OAR.results_dir

OAR.work_dir

Types

OAR.BNF

OAR.BNF

OAR.BNF

OAR.GSymbol

OAR.GSymbolSet

OAR.Grammar

OAR.ProductionRule

OAR.ProductionRuleSet

OAR.Statement

OAR.SymbolCount

OAR.Terminal

OAR.TerminalDist

Constants

Docs

Documentation for all internal names are listed below.

OAR.BNF — Type

struct BNF <: OAR.Grammar

Summary

Backus-Naur form of Grammar.

Consists of a set of terminal symbols, non-terminal symbols, and production rules.

Fields

N::Set{String}: Non-terminal symbols of the grammar.

T::Set{String}: Terminal symbols of the grammar.

S::Vector{String}: Definition of a statement in this grammar.

P::Dict{String, Set{String}}: The set of production rules of the grammar.

OAR.BNF — Method

BNF(S::Vector{String})

Summary

Constructor for a Backus-Naur Form grammer with an initial statement of non-terminal symbols.

Arguments

N::STatement: an initial set of non-terminal grammar symbols.

Method List / Definition Locations

BNF(S)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:87.

OAR.BNF — Method

BNF()

Summary

Default constructor for the Backus-Naur Form.

Method List / Definition Locations

BNF()

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:99.

OAR.GSymbol — Type

GSymbol

Description

A grammar symbol is a String.

OAR.GSymbolSet — Type

GSymbolSet

Description

A set of GSymbols.

OAR.Grammar — Type

abstract type Grammar

Summary

Abstract type for formal grammars.

Fields

OAR.ProductionRule — Type

ProductionRule

Description

A production rule is a set of GSymbols.

OAR.ProductionRuleSet — Type

ProductionRuleSet

Description

A production rule set is simply a set of ProductionRules.

OAR.Statement — Type

Statement

Description

A statement is an ordered vector of GSymbols.

OAR.SymbolCount — Type

SymbolCount

Description

The structure of the counter for symbols in a ProtoNode.

OAR.Terminal — Type

Terminal

Description

Definition of Terminal symbols used throughtout GramART.

OAR.TerminalDist — Type

TerminalDist

Description

Terminal Distribution definition that is a dictionary mapping from Terminals to probabilities.

OAR.DescretizedBNF — Method

DescretizedBNF(S::Vector{String}; bins) -> OAR.BNF

Summary

Creates a grammer for discretizing a set of symbols into a number of bins.

Arguments

N::Statement: the set of non-terminal grammar symbols to use for binning.

bins::Integer=10: optional, the granularity/number of bins.

Method List / Definition Locations

DescretizedBNF(S; bins)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:121.

OAR.is_nonterminal — Method

is_nonterminal(grammar::OAR.Grammar, symb::String) -> Any

Summary

Checks if a symbol is non-terminal in the grammar.

Method List / Definition Locations

is_nonterminal(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:175.

OAR.is_terminal — Method

is_terminal(grammar::OAR.Grammar, symb::String) -> Any

Summary

Checks if a symbol is terminal in the grammar.

Method List / Definition Locations

is_terminal(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:168.

OAR.join_gsymbol — Method

join_gsymbol(symb::String, num::Integer) -> String

Summary

Returns a new GSymbol by adding a suffix.

Method List / Definition Locations

join_gsymbol(symb, num)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:110.

OAR.parse_grammar — Method

parse_grammar(
    grammar::OAR.Grammar,
    statement::Vector{String}
)

Summary

Parses and checks that a statement is permissible under a grammer.

Method List / Definition Locations

parse_grammar(grammar, statement)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:154.

OAR.random_produce — Method

random_produce(grammar::OAR.Grammar, symb::String) -> Any

Summary

Produces a random terminal from the non-terminal using the corresponding production rule.

Method List / Definition Locations

random_produce(grammar, symb)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:161.

OAR.random_statement — Method

random_statement(grammar::OAR.Grammar) -> Vector{String}

Summary

Generates a random statement from a grammar.

Method List / Definition Locations

random_statement(grammar)

defined at /home/runner/work/OAR/OAR/src/Backus-Naur.jl:182.

OAR.results_dir — Method

results_dir(args...) -> String

Summary

Points to the results directory.

Method List / Definition Locations

results_dir(args)

defined at /home/runner/work/OAR/OAR/src/drwatson.jl:94.

OAR.work_dir — Method

work_dir(args...) -> String

Summary

Points to the work directory containing raw datasets, processed datasets, and results.

Method List / Definition Locations

work_dir(args)

defined at /home/runner/work/OAR/OAR/src/drwatson.jl:78.