Index

This page lists the core functions and types of the Julia component of the OAR project.

"Table" of Contents

Modules	Functions	Types	Constants

Index	Index	Index	Index

Docs	Docs	Docs	Docs

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

OAR.OAR

Functions

Types

OAR.GSymbol

OAR.GSymbol

OAR.ProtoNode

OAR.ProtoNode

OAR.ProtoNode

OAR.START

OAR.START

OAR.START

OAR.TreeNode

OAR.TreeNode

OAR.TreeNode

Constants

OAR.OAR_VERSION

Docs

This section lists the documentation for every exported name of the OAR package.

Modules

OAR.OAR — Module

A module encapsulating the experimental driver code for the OAR project.

Imports

The following names are imported by the package as dependencies:

AdaptiveResonance

ArgParse

Base

CSV

Clustering

Core

DataFrames

DelimitedFiles

Distributed

DocStringExtensions

DrWatson

InvertedIndices

Lerche

MLDatasets

MLUtils

NumericalTypeAliases

Parameters

Pkg

Plots

ProgressMeter

Random

Reexport

StatsPlots

Exports

The following names are exported and available when using the package:

GSymbol

OAR_VERSION

ProtoNode

START

TreeNode

Functions

Types

OAR.GSymbol — Type

struct GSymbol{T}

Summary

Definition of a grammar symbol with arbitrary datatype and a boolean flag for if the symbol is terminal or not.

Fields

data::Any: The grammar symbol of type T.

terminal::Bool: Boolean flag if the symbol is terminal (true) or nonterminal (false).

OAR.GSymbol — Method

Constructor for a grammar symbol from just the provided data (defaults to being terminal).

Arguments

data::T where T <: Any: the piece of data comprising the grammar symbol of any type.

OAR.ProtoNode — Type

struct ProtoNode <: OAR.ARTNode

Summary

ProtoNode struct, used to generate tree prototypes, which are the templates of OAR.START.

Fields

dist::Dict{GSymbol{String}, Float64}: The TerminalDist distribution over all symbols at this node.

N::Dict{GSymbol{String}, Int64}: The SymbolCount update counters for each symbol.

children::Dict{GSymbol{String}, ProtoNode}: The children of this node (Dict{STARTSymbol, ProtoNode}).

stats::OAR.ProtoNodeStats: The mutable ProtoNodeStats options and stats of the ProtoNode.

OAR.ProtoNode — Method

ProtoNode(symbols::Set{GSymbol{T}} where T) -> ProtoNode

Summary

Constructor for a zero-initialized OAR.START OAR.ProtoNode.

Arguments

symbols::SymbolSet: the terminal symbols to initialize the node with.

Method List / Definition Locations

ProtoNode(symbols)

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:179.

OAR.ProtoNode — Method

ProtoNode() -> ProtoNode

Summary

Empty constructor for a OAR.START OAR.ProtoNode.

Method List / Definition Locations

ProtoNode()

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:163.

OAR.START — Type

struct START <: OAR.SingleSTART

Summary

Definition of a START module.

Contains the ProtoNodes and CFG grammar that is used for processing statements and generating nodes.

Fields

protonodes::Vector{ProtoNode}: The OAR.ProtoNodes of the START module.

grammar::OAR.CFG: The OAR.CFG (Context-Free Grammar) used for processing data (statements).

opts::OAR.opts_START: The OAR.opts_START hyperparameters of the START module.

labels::Vector{Int64}: Incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

T::Vector{Float64}: Activation values for every weight for a given sample.

M::Vector{Float64}: Match values for every weight for a given sample.

stats::Dict{String, Any}: Dictionary of mutable statistics for the module.

OAR.START — Method

START(grammar::OAR.CFG, opts::OAR.opts_START) -> START

Summary

Constructor for an OAR.START module that takes a CFG grammar and automatically sets up the ProtoNode tree.

Arguments

grammar::CFG: the [OAR.CFG] context-free grammar to use.

opts::opts_START: a custom set of OAR.START options to use.

Method List / Definition Locations

START(grammar, opts)

defined at /home/runner/work/OAR/OAR/src/lib/start/start.jl:108.

OAR.START — Method

START(grammar::OAR.CFG; kwargs...) -> START

Summary

Constructor for an OAR.START module that takes a OAR.CFG grammar and an optional list of keyword arguments for the options.

Arguments

grammar::CFG: the [OAR.CFG] context-free grammar to use.

kwargs...: a list of keyword arguments for the OAR.opts_START options struct.

Method List / Definition Locations

START(grammar; kwargs...)

defined at /home/runner/work/OAR/OAR/src/lib/start/start.jl:131.

OAR.TreeNode — Type

struct TreeNode <: OAR.ARTNode

Summary

Tree node for a START module.

Fields

t::GSymbol{String}: The STARTSymbol symbol for the node.

children::Vector{TreeNode}: Children nodes of this node.

OAR.TreeNode — Type

TreeNode(name::AbstractString) -> TreeNode
TreeNode(
    name::AbstractString,
    is_terminal::Bool
) -> TreeNode

Summary

Constructor for a OAR.START OAR.TreeNode, taking a string name of the symbol and if it is terminal or not.

Arguments

name::AbstractString: the string name of the symbol to instantiate the OAR.TreeNode with.

is_terminal::Bool: flag for if the symbol in the node is terminal or not.

Method List / Definition Locations

TreeNode(name)
TreeNode(name, is_terminal)

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:213.

OAR.TreeNode — Method

TreeNode(symb::GSymbol{String}) -> TreeNode

Summary

Constructor for a OAR.START OAR.TreeNode taking an existing OAR.STARTSymbol.

Arguments

symb::STARTSymbol: the preconstructed STARTSymbol used for constructing the OAR.TreeNode.

Method List / Definition Locations

TreeNode(symb)

defined at /home/runner/work/OAR/OAR/src/lib/start/common.jl:199.

Constants

OAR.OAR_VERSION — Constant

A constant that contains the version of the installed OAR package.

This value is computed at compile time, so it may be used to programmatically verify the version of OAR that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.